==> 01rscripts/01preprocessing/00_control.R <==
################################################################################
## control script include alle directories to source the subproject           ##
##                                                                            ##
##                                                                            ##
## @author: Toni Lange                                                        ##
## @institution:                                                              ##
##      Center for Evidence-based Healthcare                                  ##
##      University Hospital and Medical Faculty Carl Gustav Carus,            ##
##      TU Dresden, Germany                                                   ##
##                                                                            ##
## @date: 04.09.2018                                                          ##
################################################################################

options(max.print = 500, stringsAsFactors = F, expressions = 500000, warning.length = 5000L)

#****************************************
# run in Terminal with projetc directory
#****************************************
# # # source("01rscripts/01preprocessing/00_control.R")
sink("00data/rdata/01preprocessing/preprocessing.txt")




# relative directory for source path
fd <- "01rscripts/01preprocessing/"

#*******************************************************************************
# packages
#*******************************************************************************
source(paste(fd,"01_packages.R", sep = ""))

#***********************
tic("total preprocessing")
#************************

#*******************************************************************************
# functions
#*******************************************************************************
source(paste(fd,"02_functions.R", sep = ""))

#*******************************************************************************
# import
#*******************************************************************************
source(paste(fd,"03_data_import.R", sep = ""))
# total script: 00:00:01.573 processing time

#*******************************************************************************
# remove additional content
#*******************************************************************************
source(paste(fd,"05_remove_var_content.R", sep = ""))
# total script: 00:00:06.603 processing time

#*******************************************************************************
# data preprocessing
#*******************************************************************************
source(paste(fd,"06_cleaning.R", sep = ""))
# total script: 00:03:50.317 processing time

#*******************************************************************************
# data partitioning
#*******************************************************************************
source(paste(fd,"07_data_partitioning.R", sep = ""))
# total script: 00:00:11.871 processing time

#*******************************************************************************
# data sampling
#*******************************************************************************
source(paste(fd,"08_data_sampling.R", sep = ""))
# total script: 00:00:16.18 processing time

#*******************************************************************************
# define grids
#*******************************************************************************
source(paste(fd,"09_grids_sl.R", sep = ""))
# total script: 00:00:27.808 processing time

#*******************************************************************************
# token 
#*******************************************************************************
source(paste(fd,"10_token.R", sep = ""))
#save token: 00:12:38.519 processing time

#*******************************************************************************
# dtm
#*******************************************************************************
source(paste(fd,"11_dtm.R", sep = ""))
# save dtm: 00:09:21.657 processing time
#*******************************************************************************
# split
#*******************************************************************************
source(paste(fd,"12_split_bow.R", sep = ""))

#*******************************************************************************
# split - for use with wordembedding
#*******************************************************************************
source(paste(fd,"13_split_emb.R", sep = ""))

#************************
toc()
#************************

#***********************
sink()
#***********************
