==> 01rscripts/01preprocessing/01_packages.R <==
#*******************************************************************************
#*******************************************************************************
##########################     packages      ###################################
#*******************************************************************************
#*******************************************************************************

needed.packages <- c(
  #+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  # data import
  #+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  # SQL - connect
  # "RMySQL",
  # Exel - connect
  # "readxl", "writexl",
  #+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  # data transformation
  #+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  # tidy universe
  "tidyverse", "magrittr", "rlang",
  # textmining
  "tidytext", "tm",
  # "qdap",
  "textcat", "textclean", "textstem", "SnowballC",
  #+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  # modeling
  #+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  # machine learning plattform
  "caret", "ROSE",
  #+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  # data visualization
  #+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  
  #+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  # programming
  #+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  "tictoc", "filenamer", "doSNOW", "doParallel", "pbapply"
  
)
len <- needed.packages %in% installed.packages()[, "Package"]
if (sum(!len) > 0) {
  install.packages(needed.packages[!len], dependencies = TRUE)
}
invisible(lapply(needed.packages, library, character.only = TRUE))
rm(len)
rm(needed.packages)
