==> 01rscripts/01preprocessing/03_data_import.R <==
#*******************************************************************************
#*******************************************************************************
############################     import_data      ##############################
#*******************************************************************************
#*******************************************************************************

# import raw data

#*******************************************************************************

#+++++++++++++++++++++++
tic("total script: 03_data_import.R")
#+++++++++++++++++++++++

#+++++++++++++++++++++++
tic("load data")
#+++++++++++++++++++++++

# # import via SQL
# #++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# # UTF-8 encoding
# 
# df_data <- loadAllData(table = "tbldata")
# 
# # save as tibble
# df_data <- as_tibble(df_data)
# names(df_data) <- tolower(names(df_data))
# 
# df_data <- df_data %>% select(id,rev_id, pub_id, title, abstract, rf_tas, rf_fts)

# import via RDS
# #++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# file preparation
df_data <- readRDS("00data/rdata/00df_import.RDS") 

#+++++++++++++++++++++++
toc()
#+++++++++++++++++++++++
#*******************************************************************************
########################      remove objects     ###############################
#*******************************************************************************

rm(loadAllData)

#*******************************************************************************
########################      save image     ###################################
#*******************************************************************************

#+++++++++++++++++++++++
tic("save data")
#+++++++++++++++++++++++

save.image("00data/rdata/01preprocessing/03_data_import.RData")

#+++++++++++++++++++++++
toc()
toc()
#+++++++++++++++++++++++
