==> 01rscripts/01preprocessing/07_data_partitioning.R <==
#*******************************************************************************
#*******************************************************************************
#################    data partitioning/evaluation      #########################
#*******************************************************************************
#*******************************************************************************

# create list with splittet id's for each data set
# calculate prevalence

#*******************************************************************************

#***********************
tic("total script: 07_data_partitioning.R")
#***********************

#*******************************************************************************
#     load data     ############################################################
#*******************************************************************************

#***********************
tic("load data")
#***********************

load("00data/rdata/01preprocessing/06_cleaning.RData")

#***********************
toc()
#***********************

#*******************************************************************************
#   data partitioning     ######################################################
#*******************************************************************************

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# seed for data partition
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
set.seed(42424242)
n_rev <- length(table(df_data$rev_id))

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# prepare lists
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++

ls_id <- list(
  train_id = vector(mode = "list", length = n_rev),
  test_id = vector(mode = "list", length = n_rev),
  temp_id = vector(mode = "list", length = n_rev),
  train_val_id = vector(mode = "list", length = n_rev),
  val_id = vector(mode = "list", length = n_rev)
)

#*******************************************************************************
#         y_label       ########################################################
#*******************************************************************************

df_data <- df_data %>% 
  mutate(y_label = factor(rf_tas, levels = c("inclusion", "exclusion"))) %>% 
  select(-contains("rf"))

# take care for order
df_data %<>% arrange(id) %>%
  select(contains("id"), contains("y_label"), everything())

#*******************************************************************************
#         data_split       #####################################################
#*******************************************************************************

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# training data 70%
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++

for (i in 1:n_rev) {
  ls_id$train_id[[i]] <-
    df_data %>%
    filter(rev_id == i) %>%
    pull(y_label) %>%
    createDataPartition(p = .7, list = FALSE) %>%
    as.numeric()
}

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# test data 30%
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++

for (i in 1:n_rev) {
  ls_id$test_id[[i]] <-
    df_data %>%
    filter(rev_id == i) %>%
    filter(!pub_id %in% ls_id$train_id[[i]]) %>%
    pull(pub_id)
}

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# train_val 50% and validation data 20%
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++

for (i in 1:n_rev) {
  ls_id$temp_id[[i]] <-
    df_data %>%
    filter(rev_id == i) %>%
    filter(pub_id %in% ls_id$train_id[[i]]) %>%
    pull(y_label) %>%
    createDataPartition(p = .3, list = FALSE) %>%
    as.numeric()

  ls_id$val_id[[i]] <-
    df_data %>%
    filter(rev_id == i) %>%
    filter(!pub_id %in% ls_id$test_id[[i]]) %>%
    slice(ls_id$temp_id[[i]]) %>%
    pull(pub_id)

  ls_id$train_val_id[[i]] <-
    df_data %>%
    filter(rev_id == i) %>%
    filter(!pub_id %in% ls_id$test_id[[i]]) %>%
    filter(!pub_id %in% ls_id$val_id[[i]]) %>%
    pull(pub_id)
}
# remove temp_id
ls_id$temp_id <- NULL

#*******************************************************************************
#  calculate prevalence     ####################################################
#*******************************************************************************

ls_counts_org <- tidy_calc_prev(df_data)

#*******************************************************************************
#  remove objects     ##########################################################
#*******************************************************************************

rm(list = ls()[ls() %in%
  c(
    "i"
  )])

#*******************************************************************************
#    save    ###################################################################
#*******************************************************************************

#***********************
tic("save data")
#***********************

save.image("00data/rdata/01preprocessing/07_data_partitioning.RData")

#***********************
toc() # save data: 20.006 sec elapsed
toc() # total script: 21.178 sec elapsed -minutes: 0.35
#***********************
