==> 01rscripts/01preprocessing/14_check_split_bow.R <==
#*******************************************************************************
#*******************************************************************************
#############################     check      ###################################
#*******************************************************************************
#*******************************************************************************

# check split data

#*******************************************************************************

#+++++++++++++++++++++++
tic("total script: 13_check_split.R")
#+++++++++++++++++++++++

#*******************************************************************************
########################      load data     ####################################
#*******************************************************************************

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# load data
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++

#+++++++++++++++++++++++
tic("load data")
#+++++++++++++++++++++++

load("00data/rdata/01preprocessing/09_grid_sl.RData")

#+++++++++++++++++++++++
toc()
#+++++++++++++++++++++++

#*******************************************************************************
########################      single core     ##################################
#*******************************************************************************
# 
# dat_check <- data.frame(data_id = "", data = rep(0, nrow(grid_sl)), train = 0, test = 0)
# 
# 
# tic("check count")
# pb <- txtProgressBar(max = nrow(grid_sl), style = 3, width = 100)
# for(i in 1:nrow(grid_sl)){
# temp <- readRDS(paste0("00data/rdata/01preprocessing/12_split/ls_split_", grid_sl$data_id_config[i],".RDS"))
# 
# dat_check$data_id[i] <- grid_sl$data_id_config[i]
# dat_check$data[i] <- nrow(temp$data_dtm)
# dat_check$train[i] <- nrow(temp$train_dtm)
# dat_check$test[i] <- nrow(temp$test_dtm)
# setTxtProgressBar(pb, i)
# }
# close(pb)
# toc()

# #*******************************************************************************
# ########################      save data     ###################################
# #*******************************************************************************
# 
# #+++++++++++++++++++++++
# tic("save data")
# #+++++++++++++++++++++++
# # saveRDS(dat_check, "00data/rdata/01preprocessing/13_check_ls.RDS")
# 
# #+++++++++++++++++++++++
# toc()
# toc()
# #+++++++++++++++++++++++

#*******************************************************************************
########################      multi core    ####################################
#*******************************************************************************

cores <- tidy_cores()
cl <- makeCluster(cores, outfile = "", type='SOCK')
registerDoSNOW(cl)

pb <- txtProgressBar(max = nrow(grid_sl), style = 3)
progress <- function(n) setTxtProgressBar(pb, n)
opts <- list(progress = progress)

check <- function(j){
  temp <- readRDS(paste0("00data/rdata/01preprocessing/12_split/ls_split_", grid_sl$data_id_config[j],".RDS"))
  data.frame( 
             grid_sl[j,],
             n_data = nrow(temp$data_dtm),
             n_train = nrow(temp$train_dtm),
             n_test = nrow(temp$test_dtm)
  )
}

dat_check <- foreach(
  j = 1:nrow(grid_sl),
  .options.snow = opts,
  .packages = c("tidyverse"),
  .combine = rbind
) %dopar% {
  check(j)
}
stopCluster(cl)

#*******************************************************************************
########################      save data     ###################################
#*******************************************************************************

#+++++++++++++++++++++++
tic("save data")
#+++++++++++++++++++++++
# View(dat_check)
saveRDS(dat_check, "00data/rdata/01preprocessing/13_check_ls.RDS")

#+++++++++++++++++++++++
toc()
toc()
#+++++++++++++++++++++++
