==> 01rscripts/01preprocessing/15_add_info.R <==
#*******************************************************************************
#*******************************************************************************
#########################    Additional Informations    ########################
#*******************************************************************************
#*******************************************************************************

# explore additional infos

#*******************************************************************************

#*******************************************************************************
########################      load data     ####################################
#*******************************************************************************

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# load data
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++

load("00data/rdata/01preprocessing/09_grid_sl.RData")

#*******************************************************************************
########################      Info   ###########################################
#*******************************************************************************

count_ti <- str_count(string = df_data$ti_cl, pattern =  '\\w+')
mean(count_ti)
sd(count_ti)
median(count_ti)
min(count_ti)
max(count_ti)

count_ab <- str_count(string = df_data$ab_cl, pattern =  '\\w+')
mean(count_ab)
sd(count_ab)
median(count_ab)
min(count_ab)
max(count_ab)

ls_counts_org


min(str_count(string = paste(df_data$ti_cl, df_data$ab_cl), pattern =  '\\w+'))

sort(str_count(string = paste(df_data$ti_cl, df_data$ab_cl), pattern =  '\\w+'))
#*******************************************************************************
########################      Info   ###########################################
#*******************************************************************************



count_ti <- str_count(string = df_data %>% pull(ti_cl), pattern =  '\\w+')
mean(count_ti)
sd(count_ti)

count_ab <- str_count(string = df_data %>% pull(ab_cl), pattern =  '\\w+')
mean(count_ab)
sd(count_ab)
